Ich habe die Engine von dem OpenGL renderer auf einen eigenen umgebaut, nen Shader für globale Beleuchtung und PunktLichter geschrieben. Im Moment ist der ImageHandler der Renderer, das hätte ich in ne andere Klasse packen sollen, war dafür aber zu faul. Außerdem habe ich jetzt noch für nen paar Stunden alles Kommentiert, damit alle nicht Interns von unserem Spiel (Lehrpersonen) auch verstehen was überhaupt passiert.

+++ src/main/java/Map/Wall.java
@@ -3,7 +3,7 @@ package Map;
 import Physics2D.PhysicsObject2D;
 
 public class Wall extends PhysicsObject2D {
-    public Wall(int PosX, int PosY, int WallLength, int WallHeight, int TextureID) {
+    public Wall(int PosX, int PosY, float WallLength, float WallHeight, int TextureID) {
         super(PosX, PosY, WallLength, WallHeight, TextureID);
     }
 }
